Moving filters under one single component
